delete component ListItems
