+++ src/App.tsx
@@ -1,11 +1,9 @@
-/* eslint-disable import/no-named-as-default-member */
 import React, { Suspense } from 'react';
 import './App.css';
 import { RenderRoutes } from './navigation/Router';
 import ROUTER_CONFIG from './navigation/router-config';
 
 import Header from './lib/components/Header/Header';
-// import navBarElems from './lib/components/Header/header-state';
 
 
 const App: React.FC  = () => {
